requirements.txt
django==3.2.7
djangorestframework
telebot 
requests
os